integrate the conversation list into the bottom navigation
