+++ app/src/main/kotlin/com/wire/android/ui/conversation/all/AllConversationState.kt
@@ -4,7 +4,7 @@ import com.wire.android.ui.conversation.all.model.Conversation
 import com.wire.android.ui.conversation.all.model.ConversationFolder
 import com.wire.android.ui.conversation.all.model.NewActivity
 
-data class ConversationState(
+data class AllConversationState(
     val newActivities: List<NewActivity> = emptyList(),
     val conversations: Map<ConversationFolder, List<Conversation>> = emptyMap()
 )
