--- app/src/main/kotlin/com/wire/android/ui/main/convesations/all/AllConversationsTab.kt
@@ -1,16 +1,0 @@
-package com.wire.android.ui.main.convesations.all
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.wire.android.ui.common.UnderConstructionScreen
-
-@Composable
-fun AllConversationsTab() {
-    UnderConstructionScreen(screenName = "AllConversationsTab")
-}
-
-@Preview(showBackground = false)
-@Composable
-fun AllConversationsTabPreview() {
-    AllConversationsTab()
-}
