--- app/src/main/kotlin/com/wire/android/ui/main/convesations/calls/CallsConversationsTab.kt
@@ -1,16 +1,0 @@
-package com.wire.android.ui.main.convesations.calls
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.wire.android.ui.common.UnderConstructionScreen
-
-@Composable
-fun CallConversationsTab() {
-    UnderConstructionScreen(screenName = "CallConversationsTab")
-}
-
-@Preview(showBackground = false)
-@Composable
-fun CallConversationsTabPreTab() {
-    CallConversationsTab()
-}
