--- app/src/main/kotlin/com/wire/android/ui/main/convesations/mentions/MentionsConversationsTab.kt
@@ -1,16 +1,0 @@
-package com.wire.android.ui.main.convesations.mentions
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import com.wire.android.ui.common.UnderConstructionScreen
-
-@Composable
-fun MentionsConversationsTab() {
-    UnderConstructionScreen(screenName = "MentionsConversationsTab")
-}
-
-@Preview(showBackground = false)
-@Composable
-fun MentionsConversationsTabPreTab() {
-    MentionsConversationsTab()
-}
